moved to useUrlVars
